requirements.txt
Flask==2.0.1
tensorflow==2.5.0
Pillow==8.2.0
numpy==1.19.5
